SDK-1535: Move test util class
